create add vehicle page
