+++ src/pages/private/Vehicles/index.tsx
@@ -1,100 +1,64 @@
-import ModalForm, { ModalFormProps } from '@/components/Modal/ModalForm'
 import { HttpStatusCode } from '@/constants/httpStatusCode.enum'
 import useColumnSearch from '@/hooks/useColumnSearch'
-import { useQueryDriver } from '@/queries/driver'
-import {
-  useDeleteVehiclesMutation,
-  useQueryTypeOfVehicles,
-  useQueryTypeVehiclesOwner,
-  useQueryVehicles,
-  useQueryVehiclesDetails,
-  useUpdateVehiclesMutation
-} from '@/queries/vehicle'
+import { useDeleteVehiclesMutation, useQueryVehicles } from '@/queries/vehicle'
 import { DataTypeVehicle } from '@/types/DataType'
 import { handlingTsUndefined } from '@/utils/handlingTsUndefined'
+// import { useLocalStorage } from '@/utils/localStorage/localStorageService'
 import renderWithLoading from '@/utils/renderWithLoading'
 import { PlusOutlined } from '@ant-design/icons'
 import type { TableProps } from 'antd'
-import { Button, Form, Input, InputNumber, message, Popconfirm, Select, Space, Switch, Table } from 'antd'
-import TextArea from 'antd/es/input/TextArea'
-import React, { useEffect, useState } from 'react'
+import { Button, message, Popconfirm, Space, Table } from 'antd'
+import React, { useEffect } from 'react'
+import { Link } from 'react-router-dom'
 
 const VehiclesPage: React.FC = () => {
-  const [isModalOpen, setIsModalOpen] = useState(false)
-
-  const { Option } = Select
-
-  const [selectedItem, setSelectedItem] = useState<DataTypeVehicle | null>(null)
-
-  const [lastFetchedId, setLastFetchedId] = useState<number | null>(null)
-
-  const [form] = Form.useForm()
-
   const { data, refetch: refetchVehicles, isLoading } = useQueryVehicles()
 
-  const updateMutation = useUpdateVehiclesMutation()
-
   const deleteMutaion = useDeleteVehiclesMutation()
-
-  const { data: formData, refetch } = useQueryVehiclesDetails(
-    { id: selectedItem?.id },
-    {
-      enabled: false
-    }
-  )
-
-  const { data: dataTypeOfVehicles, refetch: refetchTypeOfVehicles } = useQueryTypeOfVehicles()
-
-  const { data: dataTypeOfVehiclesOwner, refetch: refetchTypeOfVehiclesOwner } = useQueryTypeVehiclesOwner()
-
-  const { data: dataTypeDriver, refetch: refetchDriver } = useQueryDriver()
 
   const dataSource = data?.map((item: any) => ({
     ...item,
     key: item.id || item.someUniqueField
   }))
 
-  const handleEdit = (item: DataTypeVehicle) => {
-    setSelectedItem(item)
-    setIsModalOpen(true)
-  }
-
   useEffect(() => {
-    if (selectedItem?.id && selectedItem.id !== lastFetchedId) {
-      refetch()
-      refetchTypeOfVehicles()
-      refetchTypeOfVehiclesOwner()
-      refetchDriver()
-      setLastFetchedId(selectedItem.id)
-    }
-  }, [refetch, refetchTypeOfVehicles, refetchTypeOfVehiclesOwner, selectedItem, lastFetchedId, refetchDriver])
+    refetchVehicles()
+  }, [refetchVehicles])
 
-  useEffect(() => {
-    if (formData) {
-      form.setFieldsValue(formData)
-      setSelectedItem((prev) => ({
-        ...prev,
-        ...formData
-      }))
-    }
-  }, [formData, form])
+  // useEffect(() => {
+  //   if (formData) {
+  //     form.setFieldsValue(formData)
+  //     setSelectedItem(null)
+  //   }
+  // }, [formData, form])
 
-  const handleFormSubmit = async (values: any) => {
-    try {
-      const response = await updateMutation.mutateAsync({ id: selectedItem?.id, body: values })
-      if (response.status === HttpStatusCode.Ok) {
-        message.success('Update successfully')
-        refetchVehicles()
-      } else {
-        message.error('Update failed')
-      }
-    } catch (error) {
-      console.error('Error updating values:', error)
-    } finally {
-      setIsModalOpen(false)
-      setSelectedItem(null)
-    }
-  }
+  // const handleFormSubmit = async (values: any) => {
+  //   const handleResponse = (response: any, successMessage: string, errorMessage: string) => {
+  //     if (response.status === HttpStatusCode.Ok) {
+  //       message.success(successMessage)
+  //       refetchVehicles()
+  //     } else {
+  //       message.error(errorMessage)
+  //     }
+  //   }
+
+  //   try {
+  //     let response
+  //     if (mode === 'add') {
+  //       console.log(values)
+  //       response = await addMutation.mutateAsync(values)
+  //       handleResponse(response, response.message, 'Add failed')
+  //     } else if (mode === 'edit' && lastFetchedId != null) {
+  //       response = await updateMutation.mutateAsync({ id: lastFetchedId, body: values })
+  //       handleResponse(response, 'Update successfully', 'Update failed')
+  //     }
+  //   } catch (error) {
+  //     console.error('Error values:', error)
+  //   } finally {
+  //     setIsModalOpen(false)
+  //     setSelectedItem(null)
+  //   }
+  // }
 
   const handleFormDelete = async (id: number) => {
     try {
@@ -110,80 +74,86 @@ const VehiclesPage: React.FC = () => {
     }
   }
 
-  const fields: ModalFormProps<DataTypeVehicle>['fields'] = [
-    {
-      name: 'description',
-      label: 'Mô tả',
-      component: <TextArea />,
-      rules: [{ required: true, message: 'Vui lòng nhập Mô tả!' }]
-    },
-    {
-      name: 'driverId',
-      label: 'Tài xế',
-      component: (
-        <Select placeholder='Chọn tài xế' style={{ width: '100%' }}>
-          {dataTypeDriver?.map((item: any) => (
-            <Option key={item.id} value={item.id}>
-              {item.userName}
-            </Option>
-          ))}
-        </Select>
-      ),
-      rules: [{ required: true, message: 'Vui lòng chọn tài xế!' }]
-    },
-    {
-      name: 'image',
-      label: 'Link ảnh',
-      component: <Input />,
-      rules: [{ required: true, message: 'Vui lòng nhập link ảnh!' }]
-    },
-    {
-      name: 'numberSeat',
-      label: 'Số ghế ngồi',
-      component: <InputNumber style={{ width: '100%' }} />,
-      rules: [{ required: true, message: 'Vui lòng nhập số chỗ ngồi!' }]
-    },
-    {
-      name: 'licensePlate',
-      label: 'Biển số xe',
-      component: <Input />,
-      rules: [{ required: true, message: 'Vui lòng nhập Biển số xe!' }]
-    },
-    {
-      name: 'vehicleTypeId',
-      label: 'Nhà xe',
-      component: (
-        <Select placeholder='Chọn nhà xe' style={{ width: '100%' }}>
-          {dataTypeOfVehicles?.map((item: any) => (
-            <Option key={item.id} value={item.id}>
-              {item.description}
-            </Option>
-          ))}
-        </Select>
-      ),
-      rules: [{ required: true, message: 'Vui lòng chọn nhà xe!' }]
-    },
-    {
-      name: 'vehicleOwner',
-      label: 'Chủ nhà xe',
-      component: (
-        <Select placeholder='Chọn nhà xe' style={{ width: '100%' }}>
-          {dataTypeOfVehiclesOwner?.map((item: any) => (
-            <Option key={item.id} value={item.id}>
-              {item.username}
-            </Option>
-          ))}
-        </Select>
-      ),
-      rules: [{ required: true, message: 'Vui lòng chọn chủ nhà xe!' }]
-    },
-    {
-      name: 'status',
-      label: 'Trạng thái',
-      component: <Switch checkedChildren='Khả dụng' unCheckedChildren='Không khả dụng' />,
-      valuePropName: 'checked'
-    }
-  ]
+  // const openAddModal = () => {
+  //   setMode('add')
+  //   form.resetFields()
+  //   setIsModalOpen(true)
+  // }
+
+  // const fields: ModalFormProps<DataTypeVehicle>['fields'] = [
+  //   {
+  //     name: 'description',
+  //     label: 'Mô tả',
+  //     component: <TextArea />,
+  //     rules: [{ required: true, message: 'Vui lòng nhập Mô tả!' }]
+  //   },
+  //   {
+  //     name: 'driverId',
+  //     label: 'Tài xế',
+  //     component: (
+  //       <Select placeholder='Chọn tài xế' style={{ width: '100%' }}>
+  //         {dataTypeDriver?.map((item: any) => (
+  //           <Option key={item.id} value={item.id}>
+  //             {item.userName}
+  //           </Option>
+  //         ))}
+  //       </Select>
+  //     ),
+  //     rules: [{ required: true, message: 'Vui lòng chọn tài xế!' }]
+  //   },
+  //   {
+  //     name: 'image',
+  //     label: 'Link ảnh',
+  //     component: <Input />,
+  //     rules: [{ required: true, message: 'Vui lòng nhập link ảnh!' }]
+  //   },
+  //   {
+  //     name: 'numberSeat',
+  //     label: 'Số ghế ngồi',
+  //     component: <InputNumber style={{ width: '100%' }} />,
+  //     rules: [{ required: true, message: 'Vui lòng nhập số chỗ ngồi!' }]
+  //   },
+  //   {
+  //     name: 'licensePlate',
+  //     label: 'Biển số xe',
+  //     component: <Input />,
+  //     rules: [{ required: true, message: 'Vui lòng nhập Biển số xe!' }]
+  //   },
+  //   {
+  //     name: 'vehicleTypeId',
+  //     label: 'Nhà xe',
+  //     component: (
+  //       <Select placeholder='Chọn nhà xe' style={{ width: '100%' }}>
+  //         {dataTypeOfVehicles?.map((item: any) => (
+  //           <Option key={item.id} value={item.id}>
+  //             {item.description}
+  //           </Option>
+  //         ))}
+  //       </Select>
+  //     ),
+  //     rules: [{ required: true, message: 'Vui lòng chọn nhà xe!' }]
+  //   },
+  //   {
+  //     name: 'vehicleOwner',
+  //     label: 'Chủ nhà xe',
+  //     component: (
+  //       <Select placeholder='Chọn nhà xe' style={{ width: '100%' }}>
+  //         {dataTypeOfVehiclesOwner?.map((item: any) => (
+  //           <Option key={item.id} value={item.id}>
+  //             {item.username}
+  //           </Option>
+  //         ))}
+  //       </Select>
+  //     ),
+  //     rules: [{ required: true, message: 'Vui lòng chọn chủ nhà xe!' }]
+  //   },
+  //   {
+  //     name: 'status',
+  //     label: 'Trạng thái',
+  //     component: <Switch checkedChildren='Khả dụng' unCheckedChildren='Không khả dụng' />,
+  //     valuePropName: 'checked'
+  //   }
+  // ]
 
   const columns: TableProps<DataTypeVehicle>['columns'] = [
     {
@@ -227,9 +197,9 @@ const VehiclesPage: React.FC = () => {
       align: 'center',
       render: (_, record) => (
         <Space size='middle'>
-          <Button onClick={() => handleEdit(record)} type='primary'>
-            Edit
-          </Button>
+          <Link to={`edit?id=${record.id}`}>
+            <Button type='primary'>Edit</Button>
+          </Link>
           <Popconfirm
             title='Are you sure to delete this item?'
             okText='Yes'
@@ -254,19 +224,22 @@ const VehiclesPage: React.FC = () => {
         content: (
           <>
             <div style={{ display: 'flex', justifyContent: 'flex-end', marginBottom: 16 }}>
-              <Button type='primary' icon={<PlusOutlined />} ghost>
-                Thêm mới
-              </Button>
+              <Link to='add'>
+                <Button type='primary' icon={<PlusOutlined />} ghost>
+                  Thêm mới
+                </Button>
+              </Link>
             </div>
             <Table columns={columns} dataSource={dataSource} />
-            <ModalForm
+            {/* <ModalForm
+              form={form}
               isVisible={isModalOpen}
               onSubmit={handleFormSubmit}
               initialValues={selectedItem}
               fields={fields}
               setIsModalOpen={setIsModalOpen}
               setSelectedItem={setSelectedItem}
-            />
+            /> */}
           </>
         )
       })}
